change logic to use scalar refs for pruning
A ref now signals pruning and the dereferenced value indicates
if the constraint was satisfied or not.

Pruning is sticky -- once a rule flags for pruning, it carries
forward.  E.g. a rule can flag \1 and even if a subsequent rule
is false, the end result is \0, meaning prune and don't return.